# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hebrew RTL E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin or user's own)
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── WISHLIST ───────────────────────────────────────────────────────────────
...
  /loyalty/rules:
    get:
      operationId: getLoyaltyRules
      tags: [loyalty]
      summary: Get loyalty program rules
      responses:
        "200":
          description: Loyalty rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyRules"
    put:
      operationId: updateLoyaltyRules
      tags: [loyalty]
      summary: Update loyalty rules (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoyaltyRules"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltyRules"
...
  /loyalty/history/{userId}:
    get:
      operationId: getLoyaltyHistory
      tags: [loyalty]
      summary: Get user's loyalty transaction history
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Loyalty history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoyaltyTransaction"

  # ─── COUPONS ────────────────────────────────────────────────────────────────
...
  /inventory/warehouses:
    get:
      operationId: listWarehouses
      tags: [inventory]
      summary: List warehouses (admin)
      responses:
        "200":
          description: Warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [inventory]
      summary: Create warehouse (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarehouseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /inventory/stock:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List stock levels (admin)
      parameters:
        - name: warehouseId
          in: query
          schema:
            type: ["integer", "null"]
        - name: productId
          in: query
          schema:
            type: ["integer", "null"]
        - name: lowStock
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Stock levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItem"
    patch:
      operationId: adjustStock
      tags: [inventory]
      summary: Adjust stock (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustStockBody"
      responses:
        "200":
          description: Updated stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
  /inventory/stock/{productId}/history:
    get:
      operationId: getStockHistory
      tags: [inventory]
      summary: Get stock history for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockHistoryItem"

  # ─── ANALYTICS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        nameHe:
          type: string
        nameEn:
          type: ["string", "null"]
        slug:
          type: string
        descriptionHe:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        costPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        specs:
          type: object
          additionalProperties: true
        stockQuantity:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        viewsCount:
          type: integer
        salesCount:
          type: integer
        ratingAverage:
          type: number
        ratingCount:
          type: integer
        weight:
          type: ["number", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nameHe
        - slug
        - price
        - stockQuantity
        - isActive
        - isFeatured
        - viewsCount
        - salesCount
        - ratingAverage
        - ratingCount
        - images
        - tags
        - createdAt
        - updatedAt
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productSku:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - price
        - subtotal
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discount:
          type: number
        shipping:
          type: number
        tax:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        couponDiscount:
          type: number
        loyaltyPointsUsed:
          type: integer
        loyaltyPointsEarned:
          type: integer
        shippingAddress:
          type: object
          additionalProperties: true
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - status
        - items
        - subtotal
        - discount
        - shipping
        - tax
        - total
        - couponDiscount
        - loyaltyPointsUsed
        - loyaltyPointsEarned
        - createdAt
        - updatedAt
    CreateOrderBody:
      type: object
      properties:
        shippingAddress:
          type: object
          additionalProperties: true
        notes:
          type: ["string", "null"]
        loyaltyPointsToUse:
          type: integer
      required:
        - shippingAddress
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        notes:
          type: ["string", "null"]
      required:
        - status
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
        - totalPages

    # ── Wishlist ──
...
    LoyaltyRules:
      type: object
      properties:
        pointsPerShekel:
          type: number
        shekelPerPoint:
          type: number
        minRedemptionPoints:
          type: integer
        maxRedemptionPercent:
          type: number
        tierMultipliers:
          type: object
          additionalProperties: true
      required:
        - pointsPerShekel
        - shekelPerPoint
        - minRedemptionPoints
        - maxRedemptionPercent
...
    LoyaltyTransaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        points:
          type: integer
        type:
          type: string
          enum: [earned, redeemed, adjusted, expired]
        reason:
          type: string
        orderId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - points
        - type
        - reason
        - createdAt

    # ── Coupons ──
...
    Warehouse:
      type: object
      properties:
        id:
          type: integer
        nameHe:
          type: string
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameHe
        - isActive
        - createdAt
    CreateWarehouseBody:
      type: object
      properties:
        nameHe:
          type: string
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - nameHe
    StockItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        warehouseId:
          type: integer
        warehouse:
          $ref: "#/components/schemas/Warehouse"
        quantity:
          type: integer
        reservedQuantity:
          type: integer
        availableQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - product
        - warehouseId
        - warehouse
        - quantity
        - reservedQuantity
        - availableQuantity
        - lowStockThreshold
        - isLowStock
        - updatedAt
    AdjustStockBody:
      type: object
      properties:
        productId:
          type: integer
        warehouseId:
          type: integer
        quantity:
          type: integer
        reason:
          type: string
      required:
        - productId
        - warehouseId
        - quantity
        - reason
    StockHistoryItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        warehouseId:
          type: ["integer", "null"]
        quantityBefore:
          type: integer
        quantityAfter:
          type: integer
        change:
          type: integer
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - quantityBefore
        - quantityAfter
        - change
        - reason
        - createdAt

    # ── Analytics ──
...